split apart some functions for easier diff
