+++ spark2-sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuApproximatePercentile.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2022, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.nvidia.spark.rapids
+
+import org.apache.spark.sql.types._
+
+object CudfTDigest {
+  val dataType: DataType = StructType(Array(
+    StructField("centroids", ArrayType(StructType(Array(
+      StructField("mean", DataTypes.DoubleType, nullable = false),
+      StructField("weight", DataTypes.DoubleType, nullable = false)
+    )), containsNull = false)),
+    StructField("min", DataTypes.DoubleType, nullable = false),
+    StructField("max", DataTypes.DoubleType, nullable = false)
+  ))
+}
