+++ spark2-sql-plugin/src/main/scala/org/apache/spark/sql/rapids/mathExpressions.scala
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2022, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql.rapids
+
+import org.apache.spark.sql.types._
+
+object GpuFloorCeil {
+  def unboundedOutputPrecision(dt: DecimalType): Int = {
+    if (dt.scale == 0) {
+      dt.precision
+    } else {
+      dt.precision - dt.scale + 1
+    }
+  }
+}
